Outsoruce issues
