+++ src/components/FetchedIssues.js
@@ -1,112 +1,11 @@
-import { NavLink } from 'react-router-dom';
-import styled from 'styled-components/macro';
-
 import HeadingIssues from './HeadingIssues';
+import Issues from './Issues';
 
-export default function Issues({ savedIssues, togglePin }) {
+export default function FetchedIssues({ savedIssues, togglePin }) {
   return (
     <main>
       <HeadingIssues />
-
-      {savedIssues.map(savedIssue => {
-        const isPinned = savedIssues.find(
-          issueId => issueId.id === savedIssue.id
-        ).isPinned;
-
-        return (
-          <Wrapper
-            key={savedIssue.id + savedIssue.title}
-            title={savedIssue.title}
-            state={savedIssue.state}
-          >
-            <Link to={`/${savedIssue.id}`}>
-              <IssueTitle>{savedIssue.title}</IssueTitle>
-              <IssueState>{savedIssue.state}</IssueState>
-            </Link>
-            <PinButton
-              aria-label="PinButton"
-              onClick={() => {
-                togglePin(savedIssue.id, savedIssues);
-              }}
-            >
-              {isPinned ? (
-                <svg
-                  role="img"
-                  aria-label="Pin fill"
-                  width="16"
-                  height="16"
-                  fill="currentColor"
-                  viewBox="0 0 16 16"
-                >
-                  <title>Pin fill</title>
-                  <path d="M9.828.722a.5.5 0 0 1 .354.146l4.95 4.95a.5.5 0 0 1 0 .707c-.48.48-1.072.588-1.503.588-.177 0-.335-.018-.46-.039l-3.134 3.134a5.927 5.927 0 0 1 .16 1.013c.046.702-.032 1.687-.72 2.375a.5.5 0 0 1-.707 0l-2.829-2.828-3.182 3.182c-.195.195-1.219.902-1.414.707-.195-.195.512-1.22.707-1.414l3.182-3.182-2.828-2.829a.5.5 0 0 1 0-.707c.688-.688 1.673-.767 2.375-.72a5.922 5.922 0 0 1 1.013.16l3.134-3.133a2.772 2.772 0 0 1-.04-.461c0-.43.108-1.022.589-1.503a.5.5 0 0 1 .353-.146z" />
-                </svg>
-              ) : (
-                <svg
-                  role="img"
-                  aria-label="Pin empty"
-                  width="16"
-                  height="16"
-                  fill="currentColor"
-                  viewBox="0 0 16 16"
-                >
-                  <title>Pin empty</title>
-                  <path d="M9.828.722a.5.5 0 0 1 .354.146l4.95 4.95a.5.5 0 0 1 0 .707c-.48.48-1.072.588-1.503.588-.177 0-.335-.018-.46-.039l-3.134 3.134a5.927 5.927 0 0 1 .16 1.013c.046.702-.032 1.687-.72 2.375a.5.5 0 0 1-.707 0l-2.829-2.828-3.182 3.182c-.195.195-1.219.902-1.414.707-.195-.195.512-1.22.707-1.414l3.182-3.182-2.828-2.829a.5.5 0 0 1 0-.707c.688-.688 1.673-.767 2.375-.72a5.922 5.922 0 0 1 1.013.16l3.134-3.133a2.772 2.772 0 0 1-.04-.461c0-.43.108-1.022.589-1.503a.5.5 0 0 1 .353-.146zm.122 2.112v-.002.002zm0-.002v.002a.5.5 0 0 1-.122.51L6.293 6.878a.5.5 0 0 1-.511.12H5.78l-.014-.004a4.507 4.507 0 0 0-.288-.076 4.922 4.922 0 0 0-.765-.116c-.422-.028-.836.008-1.175.15l5.51 5.509c.141-.34.177-.753.149-1.175a4.924 4.924 0 0 0-.192-1.054l-.004-.013v-.001a.5.5 0 0 1 .12-.512l3.536-3.535a.5.5 0 0 1 .532-.115l.096.022c.087.017.208.034.344.034.114 0 .23-.011.343-.04L9.927 2.028c-.029.113-.04.23-.04.343a1.779 1.779 0 0 0 .062.46z" />
-                </svg>
-              )}
-            </PinButton>
-          </Wrapper>
-        );
-      })}
+      <Issues issues={savedIssues} togglePin={togglePin} />
     </main>
   );
 }
-
-const Wrapper = styled.section`
-  align-items: center;
-  display: grid;
-  grid-template-columns: 10px 1fr 52px 42px;
-  grid-template-rows: 60px;
-  margin-left: 10px;
-  margin-right: 10px;
-  :nth-of-type(2n) {
-    background-color: var(--bg-color-light);
-    border-radius: 5px;
-  }
-`;
-
-const Link = styled(NavLink)`
-  align-items: center;
-  display: inherit;
-  grid-column: 1 /4;
-  grid-template-columns: 10px 1fr 52px;
-  grid-template-rows: 60px;
-  &:active,
-  &:hover,
-  &:link,
-  &:visited {
-    color: inherit;
-    text-decoration: none;
-  }
-`;
-
-const IssueTitle = styled.p`
-  grid-column: 2/3;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-`;
-
-const IssueState = styled.div`
-  grid-column: 3;
-`;
-
-const PinButton = styled.button`
-  background: none;
-  border-radius: 50%;
-  border: none;
-  grid-column: 4/6;
-  height: 42px;
-  justify-self: end;
-  width: 42px;
-`;
