+++ src/pages/MyIssues.spec.js
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react';
-import userEvent from '@testing-library/user-event';
 import { MemoryRouter } from 'react-router-dom';
+
 import MyIssues from './MyIssues';
 
 describe('MyIssues', () => {
@@ -39,50 +39,13 @@ describe('MyIssues', () => {
     expect(heading).toBeInTheDocument();
   });
 
-  it('renders issue one', () => {
+  it('renders an issue title', () => {
     render(
       <MemoryRouter>
         <MyIssues myIssues={myIssues} />
       </MemoryRouter>
     );
-    const issueOne = screen.getByText('This is an issue title');
-    expect(issueOne).toBeInTheDocument();
-  });
-
-  it('renders issue two', () => {
-    render(
-      <MemoryRouter>
-        <MyIssues myIssues={myIssues} />
-      </MemoryRouter>
-    );
-    const issueTwo = screen.getByText('Another title');
-    expect(issueTwo).toBeInTheDocument();
-  });
-
-  it('renders the buttons', () => {
-    render(
-      <MemoryRouter>
-        <MyIssues myIssues={myIssues} />
-      </MemoryRouter>
-    );
-    const buttons = screen.getAllByRole('button');
-    expect(buttons).toHaveLength(2);
-  });
-
-  it('calls callback from PinButton', () => {
-    const callback = jest.fn();
-    render(
-      <MemoryRouter>
-        <MyIssues
-          myIssues={[
-            { id: 333, title: 'Teststring3', state: 'open', clicked: true },
-          ]}
-          togglePin={callback}
-        />
-      </MemoryRouter>
-    );
-    const PinButton = screen.getByRole('button');
-    userEvent.click(PinButton);
-    expect(callback).toHaveBeenCalled();
+    const issue = screen.getByText(/Another title/);
+    expect(issue).toBeInTheDocument();
   });
 });
